+++ examples/simple.rs
@@ -26,7 +26,7 @@ use pushrod::widget::widget::*;
 fn main() {
     let window: PistonWindow = WindowSettings::new("Pushrod Window", [800, 600])
         .opengl(OpenGL::V3_2)
-        .resizable(false)
+        .resizable(true)
         .build()
         .unwrap_or_else(|error| panic!("Failed to build PistonWindow: {}", error));
     let mut prod: Pushrod = Pushrod::new(window);
@@ -49,6 +49,12 @@ fn main() {
     base_widget.set_color([0.5, 0.5, 0.5, 1.0]);
     base_widget.on_mouse_entered(Box::new(|widget_id| {
         eprintln!("Mouse entered widget {}", widget_id);
+    }));
+    base_widget.on_mouse_moved(Box::new(|widget_id, point| {
+        eprintln!("Relative mouse move: x={} y={}", point.x, point.y);
+    }));
+    base_widget.on_mouse_down(Box::new(|widget_id, button| {
+        eprintln!("Mouse button click: {:?}", button);
     }));
     prod.widget_store.add_widget(Box::new(base_widget));
 
@@ -106,11 +112,11 @@ fn main() {
     image_widget.set_size(125, 125);
     prod.widget_store.add_widget(Box::new(image_widget));
 
-    //    let mut timer = TimerWidget::new();
-    //    timer.set_timeout(1000);
-    //    timer.set_enabled(true);
-    //    //    timer.on_timeout(Box::new(|| eprintln!("Timer.")));
-    //    prod.widget_store.add_widget(Box::new(timer));
+    let mut timer = TimerWidget::new();
+    timer.set_timeout(1000);
+    timer.set_enabled(true);
+    timer.on_timeout(Box::new(|| eprintln!("Timer.")));
+    prod.widget_store.add_widget(Box::new(timer));
 
     //    prod.add_event_listener_for_window(Box::new(ExampleListener::new()));